adding the correct copy
