+++ src/common/metrics/contact_tracing.ts
@@ -1,11 +1,16 @@
 import { COLOR_MAP } from 'common/colors';
 import { Level, LevelInfoMap } from 'common/level';
+import { levelText } from 'common/utils/chart';
+import { getLevel, Metric } from 'common/metric';
+import { formatPercent, formatInteger } from 'components/Charts/utils';
+import { Projection, TRACERS_NEEDED_PER_CASE } from 'common/models/Projection';
 
-export const METRIC_NAME = 'Contact tracing';
+export const METRIC_NAME = 'Contacts traced';
 
-const SHORT_DESCRIPTION_LOW = 'Not enough contact tracing';
-const SHORT_DESCRIPTION_MEDIUM = 'Some contact tracing';
-const SHORT_DESCRIPTION_HIGH = 'A good amount of content tracing';
+const SHORT_DESCRIPTION_LOW = 'Too many cases and too little tracing';
+const SHORT_DESCRIPTION_MEDIUM =
+  'Insufficient tracing to stop the spread of COVID';
+const SHORT_DESCRIPTION_HIGH = 'Enough tracing to help contain COVID';
 const SHORT_DESCRIPTION_UNKNOWN = 'Insufficient data to assess contact tracing';
 
 const LIMIT_LOW = 0.5;
@@ -47,3 +52,59 @@ export const CONTACT_TRACING_LEVEL_INFO_MAP: LevelInfoMap = {
     detail: SHORT_DESCRIPTION_UNKNOWN,
   },
 };
+
+export const CONTACT_TRACING_DISCLAIMER =
+  'Studies suggest that in order to successfully contain COVID,' +
+  '80% of contacts for each new COVID case must be traced within 48 hours. ' +
+  'Experts estimate that tracing this requires about 10 contact tracers per new case, ' +
+  'as well as efficient processes and fast testing.';
+
+export function contactTracingStatusText(projection: Projection) {
+  const currentContactTracers = projection.currentContactTracers;
+  const currentWeeklyAverage =
+    projection.getWeeklyAverageCaseForDay() &&
+    Math.round(projection.getWeeklyAverageCaseForDay() || 0);
+  const currentContactTracingMetric = projection.currentContactTracerMetric;
+  if (
+    currentContactTracingMetric === null ||
+    currentContactTracers === null ||
+    currentWeeklyAverage === null
+  ) {
+    return 'No testing data is available.';
+  }
+  const location = projection.locationName;
+  const level = getLevel(Metric.CONTACT_TRACING, currentContactTracingMetric);
+
+  const overview =
+    `${location} has [${formatInteger(
+      currentContactTracers,
+    )}] contact tracers.` +
+    `With an average of [${formatInteger(
+      currentWeeklyAverage,
+    )}] new daily cases, ` +
+    `we estimate ${location} needs [${formatInteger(
+      currentWeeklyAverage * TRACERS_NEEDED_PER_CASE,
+    )}] ` +
+    `contact tracing staff to trace all new cases with 48 hours, before too ` +
+    `many other people are infected.`;
+
+  const contactTracingRate = levelText(
+    level,
+    `[only ${formatPercent(currentContactTracingMetric)}]`,
+    `[${formatPercent(currentContactTracingMetric)}]`,
+    `[${formatPercent(currentContactTracingMetric)}]`,
+  );
+
+  const outcomesAtLevel = levelText(
+    level,
+    'Tracing is largely ineffective relative to these levels of case growth and does little to contain the spread of COVID.',
+    'When tracing is coupled with widely available testing, COVID can be contained without resorting to lockdowns.',
+    'When tracing is coupled with widely available testing, COVID can be contained without resorting to lockdowns.',
+  );
+
+  const details =
+    `This means that ${location} is likely able to trace, ${contactTracingRate}` +
+    ` of new COVID infection with 48 hours. ${outcomesAtLevel}`;
+
+  return `${overview} ${details}`;
+}
